Fix N² callback invocations in service subscribe pattern

+++ src/modules/settings/services/power.js
@@ -1,6 +1,7 @@
 import Host from 'stores/host';
 
 let callbackCollection = [];
+let storeSubscription = null;
 
 const reboot = () => {
 	return Host.reboot();
@@ -18,7 +19,17 @@ const handleSubscription = (properties) => {
 
 const subscribe = (callbacks) => {
 	callbackCollection = _.concat(callbackCollection, callbacks);
-	return Host.subscribeToProperties(['reboot', 'shutdown'], handleSubscription);
+	if (!storeSubscription) {
+		storeSubscription = Host.subscribeToProperties(['reboot', 'shutdown'], handleSubscription);
+	}
+
+	return () => {
+		callbackCollection = _.filter(callbackCollection, (callback) => !_.includes(callbacks, callback));
+		if (_.isEmpty(callbackCollection) && storeSubscription) {
+			storeSubscription();
+			storeSubscription = null;
+		}
+	};
 };
 
 const unsubscribe = (subsciption) => {
